move errors to api
