+++ api/errors.go
@@ -1,4 +1,4 @@
-package models
+package api
 
 type Error struct {
 	Message string
